Moved popup styles to PopupBase
